turn defaultConfig into a function mkDefaultConfig
defaultConfig is missing two fields: cAddr and cNick, which the user of
the library must specify. If the user forgets this, a run time error
will occur. Using the function mkDefaultConfig, we let the type system
ensure that the user only uses complete configs.

Because this introduces an API incompatibility, this patch also bumps
the major version number according to the cabal PVP.